Moved @tryghost/stats-service to Ghost core
fix https://linear.app/ghost/issue/ENG-2068/stats-service

- this commit migrates the code and test for the stats service
  into Ghost core so we can keep all related code together and reduce
  overhead
- as a result of this change, we can remove the package and associated
  references
