--- ghost/stats-service/.eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-    plugins: ['ghost'],
-    extends: [
-        'plugin:ghost/node'
-    ]
-};
