--- ghost/stats-service/README.md
@@ -1,11 +1,0 @@
-# Stats
-
-Stats service
-
-## Usage
-
-
-## Test
-
-- `yarn lint` run just eslint
-- `yarn test` run lint and tests
