--- ghost/stats-service/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@tryghost/stats-service",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/ghost/stats-service",
-  "private": true,
-  "author": "Ghost Foundation",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing TZ=UTC c8 --all --check-coverage --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache",
-    "posttest": "yarn lint"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "@types/luxon": "3.4.2",
-    "@types/mocha": "10.0.1",
-    "@types/sinon": "10.0.16",
-    "c8": "8.0.1",
-    "knex": "2.4.2",
-    "luxon": "3.5.0",
-    "mocha": "10.2.0",
-    "should": "13.2.3",
-    "sinon": "15.2.0",
-    "sqlite3": "5.1.7"
-  },
-  "dependencies": {
-    "moment": "2.29.3"
-  }
-}
